Removed csrf token from GET requests
Those GET requests are for retrieving data, not state-changing operations which doesn't make sense to include a CSRF token. Basically, it's unnecessary.

Additionally, for security reasons, state-changing operations should usually be in POST requests.

Furthermore, stealing of data is not possible as the GET request will fetch the data response to the victim, not the attacker.

More info:
- [Security StackExchange](https://security.stackexchange.com/questions/115794/should-i-use-csrf-protection-for-get-requests)
- [stackoverflow](https://stackoverflow.com/questions/45229789/security-against-csrf-attacks-via-get-requests)